+++ AbyssMod/Patches/PatchManager.cs
@@ -19,6 +19,7 @@ public static class PatchManager
         Harmony.CreateAndPatchAll(typeof(TranslationPatch));
         Harmony.CreateAndPatchAll(typeof(ItemPatch));
         Harmony.CreateAndPatchAll(typeof(GeneralTextPatch));
+        Harmony.CreateAndPatchAll(typeof(MasterDataTranslationPatch));
 #if DEBUG
         Harmony.CreateAndPatchAll(typeof(DebugPatch));
 #endif