+++ AbyssMod/Services/TableMapping.cs
@@ -1,0 +1,8 @@
+using System.Collections.Generic;
+
+namespace AbyssMod.Services;
+
+public class TableMapping
+{
+    public List<FieldEntry> Fields { get; } = [];
+}